Revert "TOP-18 feat: feature_signin 모듈 인증화면 구현"
