+++ .github/workflows/android_ci.yml
@@ -1,7 +1,10 @@
 name: Android CI
 
 on:
+    push:
+        branches: [ "develop" ]
     pull_request:
+        branches: [ "develop" ]
 
 jobs:
