+++ app/src/main/java/com/tht/tht/HomeActivity.kt
@@ -9,8 +9,8 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
-import tht.core.ui.base.BaseActivity
-import tht.core.ui.base.FragmentNavigator
+import com.tht.tht.base.BaseActivity
+import com.tht.tht.base.FragmentNavigator
 import com.tht.tht.binding.viewBinding
 import com.tht.tht.databinding.ActivityHomeBinding
 import dagger.hilt.android.AndroidEntryPoint
