+++ app/src/main/java/com/tht/tht/base/State.kt
@@ -1,0 +1,3 @@
+package com.tht.tht.base
+
+interface State
