--- core/ui/src/main/java/tht/core/ui/base/UiState.kt
@@ -1,3 +1,0 @@
-package tht.core.ui.base
-
-interface UiState
