--- core/ui/src/main/java/tht/core/ui/extension/AppCompatExtension.kt
@@ -1,14 +1,0 @@
-package tht.core.ui.extension
-
-import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.lifecycleScope
-import androidx.lifecycle.repeatOnLifecycle
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.launch
-
-fun AppCompatActivity.repeatOnStarted(block: suspend CoroutineScope.() -> Unit) {
-    lifecycleScope.launch {
-        lifecycle.repeatOnLifecycle(Lifecycle.State.STARTED, block)
-    }
-}
