--- core/ui/src/main/java/tht/core/ui/extension/ContextExtension.kt
@@ -1,11 +1,0 @@
-package tht.core.ui.extension
-
-import android.content.Context
-import android.view.View
-import android.view.inputmethod.InputMethodManager
-import android.widget.Toast
-
-
-fun Context.showToast(message: String) {
-    Toast.makeText(applicationContext, message, Toast.LENGTH_SHORT).show()
-}
