+++ core/ui/src/main/res/values/colors.xml
@@ -1,9 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="transparent">#00ffffff</color>
     <color name="white_f9fafa">#f9fafb</color>
     <color name="white_d7d7d7">#d7d7d7</color>
-    <color name="white_ffffff">#ffffff</color>
 
     <color name="gray_8f8f8f">#8f8f8f</color>
     <color name="gray_8d8d8d">#8d8d8d</color>
