+++ domain/src/main/java/com/tht/tht/domain/signup/usecase/RequestVerifyUseCase.kt
@@ -13,7 +13,7 @@ class RequestVerifyUseCase(
         return kotlin.runCatching {
             withContext(dispatcher) {
                 repository.requestVerify(phone, auth).also {
-                    if(it && repository.fetchSignupUser(phone) == null) { // 해당 phone 으로 진행 중인 가입 프로세스가 없다면, 새로 생성
+                    if(it) {
                         createSignupUserUseCase(phone)
                     }
                 }
