+++ feature/signin/src/main/AndroidManifest.xml
@@ -1,19 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
 
-    <application>
-        <activity
-            android:name=".auth.VerifyActivity"
-            android:exported="false"
-            android:screenOrientation="portrait" />
-        <activity
-            android:name=".auth.PhoneAuthActivity"
-            android:exported="false"
-            android:screenOrientation="portrait" />
-        <activity
-            android:name=".intro.IntroActivity"
-            android:exported="false"
-            android:screenOrientation="portrait" />
-    </application>
-
 </manifest>
