--- feature/signin/src/main/res/drawable/bg_btn_auth_enable.xml
@@ -1,12 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="314dp"
-    android:height="54dp"
-    android:viewportWidth="314"
-    android:viewportHeight="54">
-  <path
-      android:pathData="M12,0L302,0A12,12 0,0 1,314 12L314,42A12,12 0,0 1,302 54L12,54A12,12 0,0 1,0 42L0,12A12,12 0,0 1,12 0z"
-      android:fillColor="#F9CC2E"/>
-  <path
-      android:pathData="M138.47,19.43V30.29H140.6V19.43H138.47ZM136.44,24.2C136.44,21.99 134.63,20.39 132.31,20.39C129.99,20.39 128.18,21.99 128.18,24.2C128.18,26.45 129.99,28.03 132.31,28.03C134.63,28.03 136.44,26.45 136.44,24.2ZM130.25,24.2C130.24,22.97 131.13,22.23 132.31,22.23C133.46,22.23 134.35,22.97 134.35,24.2C134.35,25.46 133.46,26.19 132.31,26.19C131.13,26.19 130.24,25.46 130.25,24.2ZM132.61,29.2H130.47V34.27H141V32.56H132.61V29.2ZM143.79,25.82C146.65,25.55 148.54,24.47 149.34,23.01C150.15,24.48 152.07,25.55 154.93,25.82L155.67,24.17C152.77,23.91 151.23,22.87 150.88,21.82H155.01V20.11H143.73V21.82H147.82C147.46,22.88 145.91,23.91 143.08,24.17L143.79,25.82ZM142.47,28.09H156.26V26.38H142.47V28.09ZM149.34,28.93C146.04,28.93 144.05,29.96 144.05,31.73C144.05,33.52 146.04,34.53 149.34,34.53C152.62,34.53 154.59,33.52 154.6,31.73C154.59,29.96 152.62,28.93 149.34,28.93ZM149.34,32.88C147.25,32.88 146.2,32.52 146.21,31.73C146.2,30.95 147.25,30.58 149.34,30.57C151.39,30.58 152.47,30.95 152.48,31.73C152.47,32.52 151.39,32.88 149.34,32.88ZM168.91,25.33V19.41H166.82V34.53H168.91V27.08H171.14V25.33H168.91ZM160.34,19.65V21.67H156.91V23.37H165.8V21.67H162.45V19.65H160.34ZM161.39,24.21C159.16,24.21 157.52,25.71 157.52,27.82C157.52,29.93 159.16,31.42 161.39,31.42C163.6,31.42 165.24,29.93 165.25,27.82C165.24,25.71 163.6,24.21 161.39,24.21ZM161.39,29.68C160.32,29.69 159.55,28.99 159.56,27.82C159.55,26.66 160.32,25.94 161.39,25.93C162.45,25.94 163.21,26.66 163.23,27.82C163.21,28.99 162.45,29.69 161.39,29.68ZM182.07,34.53H184.21V19.41H182.07V34.53ZM172.38,21V22.66H177.52C177.18,26.02 175.46,28.4 171.57,30.25L172.67,31.94C178.07,29.39 179.68,25.58 179.68,21H172.38Z"
-      android:fillColor="#222222"/>
-</vector>
