--- feature/signin/src/main/res/drawable/bg_btn_login.xml
@@ -1,12 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="310dp"
-    android:height="52dp"
-    android:viewportWidth="310"
-    android:viewportHeight="52">
-  <path
-      android:pathData="M26,0L284,0A26,26 0,0 1,310 26L310,26A26,26 0,0 1,284 52L26,52A26,26 0,0 1,0 26L0,26A26,26 0,0 1,26 0z"
-      android:fillColor="#F9CC2E"/>
-  <path
-      android:pathData="M142.18,28.01H146.6V26.59H138.15V24.77H146.27V20.23H136.4V21.64H144.55V23.39H136.42V28.01H140.45V30.31H134.89V31.75H147.8V30.31H142.18V28.01ZM158.06,22.24C158.06,24.07 158.06,26.12 157.51,28.89L159.27,29.04C159.79,26.07 159.79,24.12 159.79,22.24V20.64H149.91V22.04H158.06V22.24ZM148.56,31.46H161.42V30.03H148.56V31.46ZM172.24,19.28V29.49H173.98V19.28H172.24ZM166.28,27.28C168.43,27.28 170.09,25.82 170.09,23.73C170.09,21.68 168.43,20.21 166.28,20.21C164.13,20.21 162.47,21.68 162.48,23.73C162.47,25.82 164.13,27.28 166.28,27.28ZM164.17,23.73C164.17,22.51 165.07,21.72 166.28,21.72C167.46,21.72 168.38,22.51 168.38,23.73C168.38,24.99 167.46,25.76 166.28,25.76C165.07,25.76 164.17,24.99 164.17,23.73ZM166.39,28.45H164.64V33.12H174.38V31.71H166.39V28.45Z"
-      android:fillColor="#222222"/>
-</vector>
