--- feature/signin/src/main/res/drawable/bg_btn_signup.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item>
-        <shape android:shape="rectangle">
-            <size android:width="310dp"
-                android:height="52dp"
-                />
-            <solid android:color="@color/transparent"/>
-            <corners android:radius="28dp"/>
-            <stroke android:width="1dp"
-                android:color="@color/white_f9fafa"/>
-        </shape>
-    </item>
-</selector>
