--- feature/signin/src/main/res/drawable/ic_delete.xml
@@ -1,13 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24">
-  <path
-      android:pathData="M4.929,4.929C1.024,8.834 1.024,15.166 4.929,19.071C8.834,22.976 15.166,22.976 19.071,19.071C22.976,15.166 22.976,8.834 19.071,4.929C15.166,1.024 8.834,1.024 4.929,4.929Z"
-      android:fillColor="#8D8D8D"/>
-  <path
-      android:pathData="M9.878,8.464C9.488,8.074 8.855,8.074 8.464,8.464C8.074,8.855 8.074,9.488 8.464,9.879L10.586,12L8.464,14.121C8.074,14.512 8.074,15.145 8.464,15.535C8.855,15.926 9.488,15.926 9.878,15.535L12,13.414L14.121,15.535C14.512,15.926 15.145,15.926 15.535,15.535C15.926,15.145 15.926,14.512 15.535,14.121L13.414,12L15.535,9.879C15.926,9.488 15.926,8.855 15.535,8.464C15.145,8.074 14.512,8.074 14.121,8.464L12,10.586L9.878,8.464Z"
-      android:fillColor="#161616"
-      android:fillType="evenOdd"/>
-</vector>
