--- feature/signin/src/main/res/drawable/ic_error.xml
@@ -1,13 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="16dp"
-    android:height="16dp"
-    android:viewportWidth="16"
-    android:viewportHeight="16">
-  <path
-      android:pathData="M8,2.667C5.054,2.667 2.666,5.055 2.666,8.001C2.666,10.946 5.054,13.334 8,13.334C10.945,13.334 13.333,10.946 13.333,8.001C13.333,5.055 10.945,2.667 8,2.667ZM1.333,8.001C1.333,4.319 4.318,1.334 8,1.334C11.682,1.334 14.666,4.319 14.666,8.001C14.666,11.682 11.682,14.667 8,14.667C4.318,14.667 1.333,11.682 1.333,8.001Z"
-      android:fillColor="#666666"
-      android:fillType="evenOdd"/>
-  <path
-      android:pathData="M8,4.666C7.632,4.666 7.333,4.965 7.333,5.333V8.666C7.333,9.034 7.632,9.333 8,9.333C8.368,9.333 8.666,9.034 8.666,8.666V5.333C8.666,4.965 8.368,4.666 8,4.666ZM8,9.999C7.632,9.999 7.333,10.298 7.333,10.666C7.333,11.034 7.632,11.333 8,11.333C8.368,11.333 8.666,11.034 8.666,10.666C8.666,10.298 8.368,9.999 8,9.999Z"
-      android:fillColor="#666666"/>
-</vector>
