--- feature/signin/src/main/res/drawable/ic_intro_logo.xml
@@ -1,35 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="116dp"
-    android:height="97dp"
-    android:viewportWidth="116"
-    android:viewportHeight="97">
-    <path
-        android:pathData="M111,97C107.01,97.21 97.34,89.69 93,85.91C97.25,83.79 105.46,80.26 104.35,83C103.25,85.75 108.32,93.48 111,97Z"
-        android:fillColor="#F9CC2E"/>
-    <path
-        android:pathData="M0,11.75h67.17v38.62h-67.17z"
-        android:fillColor="#F9C02E"/>
-    <path
-        android:pathData="M67.17,11.75L77.25,0V38.62L67.17,50.38V11.75Z"
-        android:fillColor="#D8B127"
-        android:fillAlpha="0.9"/>
-    <path
-        android:pathData="M105.79,50.38L115.87,38.62V77.25L105.79,89V50.38Z"
-        android:fillColor="#D8B127"
-        android:fillAlpha="0.9"/>
-    <path
-        android:pathData="M77.25,0h38.62v38.62h-38.62z"
-        android:fillColor="#F9CC2E"/>
-    <path
-        android:pathData="M0,50.38h105.79v38.62h-105.79z"
-        android:fillColor="#F9CC2E"/>
-    <path
-        android:pathData="M8.25,26.37H3.95V23.65H15.84V26.37H11.54V37.79H8.25V26.37ZM30.32,30.72C30.32,32.23 30.03,33.53 29.44,34.63C28.86,35.73 28.05,36.56 27.04,37.13C26.02,37.7 24.87,37.99 23.6,37.99C22.32,37.99 21.17,37.7 20.15,37.13C19.13,36.55 18.33,35.72 17.75,34.62C17.16,33.53 16.87,32.23 16.87,30.72C16.87,29.21 17.16,27.9 17.74,26.81C18.33,25.72 19.13,24.88 20.14,24.31C21.17,23.74 22.32,23.46 23.6,23.46C24.87,23.46 26.02,23.74 27.04,24.31C28.05,24.88 28.86,25.72 29.44,26.81C30.03,27.91 30.32,29.21 30.32,30.72ZM26.93,30.72C26.93,29.8 26.79,29.01 26.53,28.36C26.26,27.71 25.88,27.22 25.38,26.89C24.88,26.55 24.29,26.38 23.6,26.38C22.91,26.38 22.31,26.55 21.82,26.89C21.32,27.22 20.94,27.71 20.67,28.36C20.41,29.01 20.28,29.8 20.28,30.72C20.28,31.65 20.41,32.43 20.67,33.08C20.94,33.73 21.32,34.22 21.82,34.56C22.31,34.89 22.91,35.06 23.6,35.06C24.29,35.06 24.88,34.89 25.38,34.56C25.88,34.22 26.26,33.73 26.53,33.08C26.79,32.43 26.93,31.65 26.93,30.72ZM38.27,23.65C39.32,23.65 40.23,23.85 41,24.26C41.78,24.67 42.37,25.25 42.79,25.98C43.21,26.72 43.42,27.57 43.42,28.52C43.42,29.49 43.21,30.34 42.78,31.07C42.36,31.8 41.75,32.36 40.96,32.76C40.17,33.17 39.24,33.37 38.18,33.37H35.81V37.79H32.49V23.65H38.27ZM37.57,30.73C38.08,30.73 38.51,30.64 38.87,30.46C39.23,30.27 39.5,30.02 39.69,29.69C39.87,29.36 39.96,28.97 39.96,28.52C39.96,27.85 39.76,27.32 39.34,26.93C38.93,26.54 38.34,26.35 37.57,26.35H35.81V30.73H37.57ZM45.28,23.65H48.61V37.79H45.28V23.65ZM60.34,28.75C60.28,28.26 60.13,27.84 59.89,27.49C59.64,27.13 59.32,26.86 58.92,26.67C58.52,26.48 58.06,26.38 57.54,26.38C56.84,26.38 56.24,26.55 55.74,26.89C55.24,27.24 54.85,27.73 54.58,28.39C54.31,29.04 54.18,29.81 54.18,30.72C54.18,31.66 54.31,32.45 54.58,33.1C54.86,33.75 55.24,34.24 55.74,34.57C56.24,34.89 56.83,35.06 57.52,35.06C58.28,35.06 58.91,34.86 59.41,34.48C59.91,34.09 60.22,33.54 60.34,32.85L63.71,32.87C63.62,33.74 63.32,34.57 62.8,35.35C62.28,36.14 61.57,36.77 60.65,37.26C59.74,37.74 58.68,37.99 57.46,37.99C56.17,37.99 55.02,37.7 54.01,37.13C53,36.56 52.21,35.73 51.64,34.63C51.06,33.53 50.77,32.23 50.77,30.72C50.77,29.21 51.06,27.91 51.64,26.81C52.23,25.72 53.03,24.88 54.04,24.31C55.05,23.74 56.19,23.46 57.46,23.46C58.59,23.46 59.61,23.66 60.51,24.08C61.41,24.5 62.14,25.1 62.7,25.9C63.26,26.69 63.59,27.64 63.71,28.75H60.34Z"
-        android:fillColor="#161616"/>
-    <path
-        android:pathData="M82.73,25H79.76V13.69H82.73V18.14H86.91V13.69H89.89V25H86.91V20.55H82.73V25ZM102.44,19.34C102.44,20.56 102.2,21.61 101.72,22.49C101.24,23.36 100.59,24.03 99.76,24.48C98.93,24.93 98,25.15 96.97,25.15C95.93,25.15 94.99,24.93 94.17,24.48C93.34,24.02 92.68,23.35 92.2,22.48C91.73,21.6 91.49,20.56 91.49,19.34C91.49,18.12 91.73,17.07 92.2,16.2C92.68,15.32 93.33,14.66 94.16,14.21C94.99,13.76 95.92,13.53 96.97,13.53C98,13.53 98.93,13.76 99.76,14.21C100.59,14.66 101.24,15.32 101.72,16.2C102.2,17.08 102.44,18.12 102.44,19.34ZM99.38,19.34C99.38,18.3 99.17,17.51 98.76,16.96C98.34,16.4 97.74,16.13 96.97,16.13C96.19,16.13 95.59,16.4 95.18,16.96C94.76,17.51 94.55,18.3 94.55,19.34C94.55,20.38 94.76,21.18 95.18,21.73C95.59,22.28 96.19,22.56 96.97,22.56C97.74,22.56 98.34,22.28 98.76,21.73C99.17,21.18 99.38,20.38 99.38,19.34ZM106.6,16.09H103.19V13.69H112.95V16.09H109.54V25H106.6V16.09Z"
-        android:fillColor="#161616"/>
-    <path
-        android:pathData="M9.38,66.44H4.65V63.44H17.73V66.44H13V79H9.38V66.44ZM33.66,71.22C33.66,72.88 33.34,74.32 32.69,75.52C32.05,76.73 31.17,77.64 30.05,78.27C28.93,78.9 27.67,79.21 26.26,79.21C24.86,79.21 23.59,78.9 22.47,78.27C21.35,77.64 20.47,76.72 19.83,75.51C19.19,74.31 18.86,72.88 18.86,71.22C18.86,69.55 19.18,68.12 19.82,66.92C20.46,65.72 21.35,64.8 22.46,64.17C23.59,63.54 24.86,63.23 26.26,63.23C27.67,63.23 28.93,63.54 30.05,64.17C31.17,64.8 32.05,65.72 32.69,66.92C33.34,68.13 33.66,69.56 33.66,71.22ZM29.92,71.22C29.92,70.2 29.78,69.34 29.48,68.62C29.19,67.91 28.77,67.37 28.22,67C27.68,66.64 27.02,66.45 26.26,66.45C25.5,66.45 24.85,66.64 24.3,67C23.76,67.37 23.34,67.91 23.04,68.62C22.75,69.34 22.61,70.2 22.61,71.22C22.61,72.24 22.75,73.11 23.04,73.82C23.34,74.53 23.76,75.07 24.3,75.44C24.85,75.81 25.5,75.99 26.26,75.99C27.02,75.99 27.68,75.81 28.22,75.44C28.77,75.07 29.19,74.53 29.48,73.82C29.78,73.11 29.92,72.24 29.92,71.22ZM41.74,63.44C43.31,63.44 44.67,63.75 45.82,64.38C46.98,64.99 47.86,65.89 48.48,67.06C49.1,68.22 49.41,69.61 49.41,71.21C49.41,72.82 49.1,74.21 48.48,75.38C47.86,76.55 46.98,77.44 45.82,78.07C44.67,78.69 43.31,79 41.74,79H36.04V63.44H41.74ZM41.6,75.87C42.52,75.87 43.28,75.72 43.89,75.41C44.49,75.09 44.96,74.59 45.27,73.91C45.58,73.22 45.74,72.32 45.74,71.21C45.74,70.1 45.58,69.21 45.26,68.52C44.95,67.84 44.48,67.34 43.87,67.04C43.26,66.73 42.49,66.57 41.55,66.57H39.7V75.87H41.6ZM65.57,79H61.64L60.59,75.68H55.16L54.11,79H50.18L55.45,63.44H60.3L65.57,79ZM57.81,67.29L56.05,72.88H59.7L57.93,67.29H57.81ZM70.73,73.74L65.04,63.44H69.14L72.48,69.98H72.62L75.96,63.44H80.06L74.37,73.74V79H70.73V73.74ZM82.5,63.44H85.9L83.98,69.65H81.48L82.5,63.44ZM96.55,68.08C96.5,67.5 96.26,67.05 95.85,66.74C95.44,66.43 94.86,66.27 94.12,66.27C93.64,66.27 93.23,66.33 92.89,66.47C92.55,66.59 92.3,66.77 92.13,67C91.96,67.23 91.87,67.49 91.87,67.79C91.85,68.24 92.04,68.61 92.44,68.87C92.84,69.14 93.42,69.36 94.19,69.53L95.53,69.85C97.06,70.19 98.24,70.74 99.05,71.51C99.86,72.27 100.27,73.27 100.28,74.5C100.27,75.47 100.02,76.32 99.53,77.02C99.05,77.73 98.34,78.27 97.42,78.64C96.5,79.02 95.4,79.21 94.12,79.21C92.81,79.21 91.68,79.01 90.72,78.61C89.76,78.22 89.02,77.62 88.5,76.83C87.97,76.03 87.69,75.05 87.67,73.88H91.19C91.24,74.62 91.52,75.18 92.03,75.57C92.54,75.96 93.22,76.16 94.08,76.16C94.59,76.16 95.02,76.09 95.39,75.95C95.77,75.81 96.05,75.61 96.25,75.36C96.45,75.11 96.56,74.82 96.56,74.49C96.56,74.19 96.46,73.93 96.28,73.72C96.1,73.5 95.83,73.32 95.46,73.16C95.1,73 94.62,72.85 94.03,72.72L92.4,72.34C89.51,71.66 88.07,70.23 88.08,68.04C88.07,67.09 88.33,66.26 88.84,65.53C89.36,64.8 90.08,64.24 91,63.84C91.92,63.43 92.97,63.23 94.14,63.23C95.32,63.23 96.36,63.43 97.25,63.84C98.14,64.24 98.83,64.81 99.31,65.55C99.8,66.28 100.06,67.12 100.07,68.08H96.55Z"
-        android:fillColor="#161616"/>
-</vector>
