--- feature/signin/src/main/res/drawable/ic_right_arrow.xml
@@ -1,13 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="26dp"
-    android:height="26dp"
-    android:viewportWidth="26"
-    android:viewportHeight="26">
-  <path
-      android:pathData="M16.25,19.5L9.75,13L16.25,6.5"
-      android:strokeLineJoin="round"
-      android:strokeWidth="2"
-      android:fillColor="#00000000"
-      android:strokeColor="#F9FAFA"
-      android:strokeLineCap="round"/>
-</vector>
