--- feature/signin/src/main/res/drawable/selector_btn_auth.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:state_enabled="false"
-        android:drawable="@drawable/bg_btn_auth_disenable" />
-    <item android:state_enabled="true"
-        android:drawable="@drawable/bg_btn_auth_enable" />
-</selector>
