--- feature/signin/src/main/res/layout/activity_intro.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".intro.IntroActivity"
-    android:paddingHorizontal="40dp"
-    android:paddingBottom="16dp"
-    android:background="@color/black_1e1e1e"
-    >
-
-    <ImageView
-        android:layout_marginTop="50dp"
-        android:src="@drawable/ic_intro_logo"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toTopOf="@id/btn_signup"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        />
-
-    <androidx.appcompat.widget.AppCompatButton
-        style="@style/SubT1SemiBold"
-        android:id="@+id/btn_signup"
-        android:textColor="@color/white_f9fafa"
-        android:gravity="center"
-        android:layout_marginBottom="16dp"
-        app:layout_constraintBottom_toTopOf="@id/btn_login"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:background="@drawable/bg_btn_signup"
-        android:foreground="?android:attr/selectableItemBackground"
-        android:text="@string/signup"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        />
-
-    <androidx.appcompat.widget.AppCompatButton
-        android:id="@+id/btn_login"
-        android:layout_marginBottom="16dp"
-        app:layout_constraintBottom_toTopOf="@id/tv_login_issue"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:background="@drawable/bg_btn_login"
-        android:foreground="?android:attr/selectableItemBackground"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        />
-
-    <TextView
-        style="@style/P2MediumTextStyle"
-        android:textColor="@color/white_f9fafa"
-        android:text="@string/message_login_issue"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:id="@+id/tv_login_issue"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
