--- feature/signin/src/main/res/layout/activity_phone_auth.xml
@@ -1,103 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".auth.PhoneAuthActivity"
-    android:background="@color/black_1e1e1e"
-    android:id="@+id/layout_background"
-    >
-
-    <include
-        android:id="@+id/item_signup_tool_bar"
-        layout="@layout/item_signup_tool_bar" />
-
-    <TextView
-        style="@style/H1BoldTextStyle"
-        android:layout_marginTop="90dp"
-        android:id="@+id/tv_phone_auth"
-        android:text="@string/phone_auth"
-        android:textColor="@color/white_f9fafa"
-        android:layout_marginHorizontal="38dp"
-        app:layout_constraintTop_toBottomOf="@id/item_signup_tool_bar"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        />
-
-    <com.google.android.material.textfield.TextInputLayout
-        android:id="@+id/layout_et_phone"
-        app:layout_constraintTop_toBottomOf="@id/tv_phone_auth"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:layout_marginTop="32dp"
-        android:layout_marginHorizontal="@dimen/AuthTextInputEditLayoutHorizontalMargin"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        style="@style/PhoneAuthTextInputLayoutStyle"
-        android:theme="@style/PhoneAuthTextInputLayoutStyle"
-        >
-        <com.google.android.material.textfield.TextInputEditText
-            style="@style/PhoneAuthInputEditTextFontStyle"
-            android:id="@+id/et_phone"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:inputType="number"
-            android:maxLines="1"
-            android:layout_marginBottom="@dimen/TextInputEditTextBottomMinusMargin"
-            />
-    </com.google.android.material.textfield.TextInputLayout>
-
-    <ImageView
-        android:layout_marginTop="1dp"
-        android:layout_marginStart="38dp"
-        android:id="@+id/iv_error"
-        app:layout_constraintTop_toBottomOf="@id/layout_et_phone"
-        app:layout_constraintStart_toStartOf="parent"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/ic_error"
-        />
-
-    <TextView
-        style="@style/Caption1MediumGrayTextStyle"
-        android:textColor="@color/gray_666666"
-        android:id="@+id/tv_phone_auth_message"
-        android:text="@string/message_phone_auth"
-        android:layout_marginStart="7dp"
-        android:layout_marginEnd="38dp"
-        app:layout_constraintTop_toTopOf="@id/iv_error"
-        app:layout_constraintStart_toEndOf="@id/iv_error"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        />
-
-    <androidx.appcompat.widget.AppCompatButton
-        android:id="@+id/btn_auth"
-        android:enabled="false"
-        android:layout_marginTop="44dp"
-        android:layout_marginHorizontal="38dp"
-        app:layout_constraintTop_toBottomOf="@id/tv_phone_auth_message"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:background="@drawable/selector_btn_auth"
-        android:foreground="?attr/selectableItemBackground"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        />
-
-    <ProgressBar
-        android:visibility="gone"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        android:id="@+id/progress"
-        android:indeterminateTint="@color/red_ef4444"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        />
-</androidx.constraintlayout.widget.ConstraintLayout>
