--- feature/signin/src/main/res/layout/item_signup_tool_bar.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.appcompat.widget.Toolbar
-    android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:background="@color/black_1e1e1e"
-    android:id="@+id/tool_bar"
-    android:layout_width="match_parent"
-    android:layout_height="?attr/actionBarSize"
-    />
-
-
