--- feature/signin/src/main/res/values/strings.xml
@@ -1,16 +1,0 @@
-<resources>
-    <string name="signup">회원가입</string>
-    <string name="login">로그인</string>
-
-    <string name="message_login_issue"><u>로그인에 어려움이 있으신가요?</u></string>
-
-    <string name="phone_auth">핸드폰 번호 인증</string>
-    <string name="hint_phone_input">01012345678</string>
-    <string name="message_phone_auth">핸드폰 번호 입력 시, 인증 코드를 문자 메세지로 발송합니다.\n핸드폰 번호는 다른 사람들에게 공유되거나 프로필에 표시하지않습니다.</string>
-    <string name="message_phone_input_error">핸드폰 번호를 다시 확인해 주세요</string>
-
-    <string name="title_verify">인증 코드 입력</string>
-    <string name="message_verify">%1s 으로\n전송된 코드를 입력해주세요</string>
-    <string name="message_verify_error">인증 코드를 다시 확인 해 주세요</string>
-    <string name="message_resend_auth"><u>인증 코드 다시 받기</u></string>
-</resources>
